Refactor: Created MarketData folder and moved hooks inside
