+++ app/src/hooks/MarketData/use24hsVolume.ts
@@ -3,8 +3,8 @@ import { BigNumber } from 'ethers/utils'
 import gql from 'graphql-tag'
 import { useEffect, useState } from 'react'
 
-import { useConnectedWeb3Context } from '../contexts'
-import { getLogger } from '../util/logger'
+import { useConnectedWeb3Context } from '../../contexts'
+import { getLogger } from '../../util/logger'
 
 const logger = getLogger('Market::use24hsVolume')
