+++ app/src/hooks/MarketData/useBlockchainMarketMakerData.tsx
@@ -2,15 +2,14 @@ import Big from 'big.js'
 import { BigNumber, bigNumberify } from 'ethers/utils'
 import { useCallback, useEffect, useState } from 'react'
 
-import { useConnectedWeb3Context } from '../contexts'
-import { ERC20Service, MarketMakerService, OracleService } from '../services'
-import { getLogger } from '../util/logger'
-import { getArbitratorFromAddress } from '../util/networks'
-import { isScalarMarket, promiseProps } from '../util/tools'
-import { BalanceItem, MarketMakerData, Status, Token } from '../util/types'
-
-import { GraphMarketMakerData } from './Graph/useGraphMarketMakerData'
-import { useContracts } from './useContracts'
+import { useConnectedWeb3Context } from '../../contexts'
+import { ERC20Service, MarketMakerService, OracleService } from '../../services'
+import { getLogger } from '../../util/logger'
+import { getArbitratorFromAddress } from '../../util/networks'
+import { isScalarMarket, promiseProps } from '../../util/tools'
+import { BalanceItem, MarketMakerData, Status, Token } from '../../util/types'
+import { GraphMarketMakerData } from '../Graph/useGraphMarketMakerData'
+import { useContracts } from '../useContracts'
 
 const logger = getLogger('useBlockchainMarketMakerData')
