+++ app/src/hooks/index.ts
@@ -1,7 +1,7 @@
 export * from './useBalance'
-export { use24hsVolume } from './use24hsVolume'
+export { use24hsVolume } from './MarketData/use24hsVolume'
 export { useAsyncDerivedValue } from './useAsyncDerivedValue'
-export { useBlockchainMarketMakerData } from './useBlockchainMarketMakerData'
+export { useBlockchainMarketMakerData } from './MarketData/useBlockchainMarketMakerData'
 export { useCheckContractExists } from './useCheckContractExists'
 export { useCollateral } from './useCollateral'
 export { useCollateralBalance } from './useCollateralBalance'
@@ -16,7 +16,7 @@ export { useFundingBalance } from './useFundingBalance'
 export { useGraphMarketMakerData } from './Graph/useGraphMarketMakerData'
 export { useGraphMarketUserTxData } from './Graph/useGraphMarketUserTxData'
 export { useGraphMeta } from './Graph/useGraphMeta'
-export { useMarketMakerData } from './useMarketMakerData'
+export { useMarketMakerData } from './MarketData/useMarketMakerData'
 export { usePolling } from './usePolling'
 export { useTokens } from './useTokens'
 export { useRealityLink } from './useRealityLink'
